add ticket_id fixes and ticket_handle mapping (#726)
* maintain a mapping from ticket * amount to ticket_handle inside contracts
* move to using int32 as ticket_handle

+++ tests/e2e/ticket_handle.ml
@@ -3,7 +3,7 @@ open Cmdliner
 let create_handle =
   let open Core_deku in
   let open Contracts in
-  let arg = [%to_yojson: bytes] (Int64.zero |> Context.Ticket_handle.to_bytes) in
+  let arg = [%to_yojson: bytes] (Int32.zero |> Context.Ticket_handle.to_bytes) in
   Format.printf "%s\n%!" (Yojson.Safe.to_string arg)
 
 let args =
